orgs/blocknotify-data-agnostic/requirements.txt
certifi==2023.11.17
charset-normalizer==2.0.12
configobj==5.0.8
dnspython==2.4.2
ecdsa==0.18.0
ECPy==1.2.5
idna==3.6
pycryptodome==3.20.0
pycurl==7.45.2
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.27.1
six==1.16.0
slick-bitcoinrpc==0.1.4
ujson==5.9.0
urllib3==1.26.18
